feat: CRUD de notas com auth, tema dark e correcoes

+++ app/Http/Controllers/PostController.php
@@ -10,8 +10,8 @@ class PostController extends Controller
     // 1. Mostrar a lista de posts
     public function index()
     {
-        $posts = Post::all();
-        return view('posts.index', compact('posts'));
+        $posts = auth()->user()->posts; 
+    return view('posts.index', compact('posts'));
     }
 
     public function create()
@@ -19,34 +19,49 @@ class PostController extends Controller
         return view('posts.create');
     }
 
-    public function store(Request $request)
-    {
-        $request->validate([
-            'titulo' => 'required',
-            'conteudo' => 'required',
-        ]);
+   public function store(Request $request)
+{
+    $request->validate([
+        'titulo' => 'required',
+        'conteudo' => 'required',
+    ]);
 
-        Post::create($request->all());
+        auth()->user()->posts()->create([
+        'titulo' => $request->titulo,
+        'conteudo' => $request->conteudo,
+    ]);
 
-        //redirecionamento
-        return redirect()->route('posts.index')->with('success', 'Nota criada com sucesso! ✨');
-    }
+    return redirect()->route('posts.index')->with('success', 'Nota criada!');
+}
 
     public function destroy(Post $post)
-    {
-        $post->delete();
-        return redirect()->route('posts.index')->with('success', 'Nota excluída permanentemente! 🗑️');
+{
+    if ($post->user_id !== auth()->id()) {
+        abort(403);
     }
+
+    $post->delete();
+
+    return redirect()->route('posts.index');
+}
 
 
 
     public function edit(Post $post)
 {
+        if ($post->user_id !== auth()->id()) {
+        abort(403, 'Acesso não autorizado.');
+    }
+
     return view('posts.edit', compact('post'));
 }
 
 public function update(Request $request, Post $post)
 {
+    if ($post->user_id !== auth()->id()) {
+        abort(403);
+    }
+
     $request->validate([
         'titulo' => 'required',
         'conteudo' => 'required',
@@ -54,7 +69,18 @@ public function update(Request $request, Post $post)
 
     $post->update($request->all());
 
-    return redirect()->route('posts.index')->with('success', 'Nota atualizada com sucesso! 📝');
+    return redirect()->route('posts.index');
 }
+
+public function show(Post $post)
+{
+    
+    if ($post->user_id !== auth()->id()) {
+        abort(403);
+    }
+
+    return view('posts.show', compact('post'));
+}
+
 
 }
